data_processing/requirements.txt
altair==5.4.0
aniso8601==9.0.1
apispec==6.4.0
apispec-webframeworks==1.0.0
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
bidict==0.23.1
binaryornot==0.4.4
blinker==1.8.2
boto3==1.34.34
botocore==1.34.161
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cookiecutter==2.5.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
deepdiff==6.7.1
dnspython==2.6.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.1
Faker==27.0.0
fastjsonschema==2.20.0
favicon==0.7.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
fonttools==4.53.1
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gevent==23.9.1
gevent-websocket==0.10.1
gitdb==4.0.11
gitignore_parser==0.1.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
htbuilder==0.6.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kthread==0.2.3
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.4.0
mysql-connector-python==9.0.0
narwhals==1.4.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openpyxl==3.1.2
ordered-set==4.1.0
packaging==24.1
pandas==2.2.0
parso==0.8.4
passlib==1.7.4
patsy==0.5.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
plotly-express==0.4.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==15.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.9
pymongo==4.6.1
PyMySQL==1.1.1
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.3
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
scipy==1.14.0
seaborn==0.13.2
setuptools==72.2.0
shapely==2.0.5
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.25
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.37.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.7
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.3.13
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
text-unidecode==1.3
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
Twisted==23.10.0
twisted-iocpsupport==1.0.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
validators==0.33.0
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
xyzservices==2024.6.0
zope.event==5.0
zope.interface==7.0.1
